+++ demos/demo1/main.js
@@ -10,12 +10,12 @@ class Task {
 	}
 
 	get_elem() {
-		let task = dom.make_element('task-container');
+		let task = dom.elem('task-container');
 		task.add_children([
-			dom.make_element('title').text(this.name),
-			dom.make_element('description').text(this.description),
+			dom.elem('title').text(this.name),
+			dom.elem('description').text(this.description),
 		]);
-		task.add_child(dom.make_element("button").text("Delete").event("click", () => {
+		task.add_child(dom.elem("button").text("Delete").event("click", () => {
 			this.parent.remove_child(task);
 		}));
 		return task.get_elem();
@@ -26,12 +26,12 @@ class Task {
 
 dom.on_page_load(() => {
 
-	let tasks_elem = dom.div("tasks-container").id("tasks");
+	let tasks_elem = dom.elem("tasks");
 	dom.get_body().add_children([
-		dom.make_element("header").text("Tasks"),
-		dom.make_element("input").id("task_name"),
-		dom.make_element("input").id("task_description"),
-		dom.make_element("button").text("Add Task").event("click", () => {
+		dom.elem("header").text("Tasks"),
+		dom.elem("input").id("task_name"),
+		dom.elem("input").id("task_description"),
+		dom.elem("button").text("Add Task").event("click", () => {
 			let name = dom.get_element_by_id("task_name").get_value();
 			let description = dom.get_element_by_id("task_description").get_value();
 			let task = new Task(name, description, tasks_elem);